Fix up pregnancy system
