+++ OpenNefia.Content/Enchantments/Impl/Parameterized/EncAmmoComponent.cs
