+++ OpenNefia.Content/Enchantments/Impl/Parameterized/EncElementalDamageComponent.cs
