+++ OpenNefia.Content/Enchantments/Impl/Parameterized/EncInvokeSpellComponent.cs
