+++ OpenNefia.Content/Enchantments/Impl/Parameterized/EncModifyResistanceComponent.cs
