+++ OpenNefia.Content/RandomGen/CharaGenSystem.cs
@@ -24,6 +24,7 @@ namespace OpenNefia.Content.RandomGen
         PrototypeId<EntityPrototype> PickRandomCharaId(EntityGenArgSet args, int minLevel = 1, PrototypeId<TagPrototype>[]? tags = null, string? fltselect = null,
             PrototypeId<RacePrototype>? raceFilter = null, string? category = null);
 
+        // TODO TryGenerateChara
         EntityUid? GenerateChara(MapCoordinates coords, PrototypeId<EntityPrototype>? id = null,
             int minLevel = 1, PrototypeId<TagPrototype>[]? tags = null, string? fltselect = null,
             PrototypeId<RacePrototype>? raceFilter = null, Quality? quality = null, EntityGenArgSet? args = null);
