--- OpenNefia.Content/Resources/Locale/en_US/Elona/Pregnant.lua
@@ -1,5 +1,0 @@
-Elona.Pregnant = {
-    SpitsAlienChildren = function(entity)
-        return ("%s spit%s alien children from %s body!"):format(_.name(entity), _.s(entity), _.his(entity))
-    end,
-}
