+++ OpenNefia.Content/Resources/Locale/ja_JP/Elona/Pregnancy.lua
@@ -1,0 +1,28 @@
+Elona.Pregnancy = {
+    Name = {
+        ChildOf = function(entityName)
+            return ("%sの子供"):format(entityName)
+        end,
+        AlienKid = function(basename)
+            return ("%sの子供"):format(basename)
+        end,
+    },
+    ButPukesOut = "しかしすぐに吐き出した。",
+    Impregnated = function(entity)
+        return ("%sは寄生された。"):format(_.name(entity))
+    end,
+
+    PatsStomach = function(entity)
+        return ("%sは不安げに腹を押さえた。"):format(_.name(entity))
+    end,
+    SomethingBreaksOut = function(entity)
+        return ("何かが%sの腹を破り飛び出した！"):format(_.name(entity))
+    end,
+
+    AlienChildrenMelt = function(entity)
+        return ("%sの体内のエイリアンは溶けた。"):format(_.name(entity))
+    end,
+    SpitsAlienChildren = function(entity)
+        return ("%sは体内のエイリアンを吐き出した！"):format(_.name(entity))
+    end,
+}
