--- OpenNefia.Content/Resources/Locale/ja_JP/Elona/Pregnant.lua
@@ -1,5 +1,0 @@
-Elona.Pregnant = {
-    SpitsAlienChildren = function(entity)
-        return ("%sは体内のエイリアンを吐き出した！"):format(_.name(entity))
-    end,
-}
